drivers: sensor: lis2dux12: store the full FIFO level in the buffer
The FIFO level read from FIFO_STATUS2 is 8 bits, 0 to 128 words, but
the fifo_count field of the FIFO buffer header had 7 bits. A full FIFO
of 128 words was recorded as 0 words, so the decoder ignored the 896
bytes read from the FIFO. Widen the field to 8 bits and shrink
reserved_1 to 4 bits, which keeps the header size.

The decoder counted the accelerometer readings of a FIFO buffer in a
uint8_t, which 128 words in 2x mode (256 readings) would now overflow:
count them in a uint16_t.

Built for native_sim with CONFIG_SENSOR_ASYNC_API. Not tested on
hardware.

+++ drivers/sensor/st/lis2dux12/lis2dux12_decoder.c
@@ -164,7 +164,7 @@ static int lis2dux12_decoder_get_frame_count(const uint8_t *buffer,
 	const struct lis2dux12_fifo_data *edata = (const struct lis2dux12_fifo_data *)buffer;
 	const uint8_t *buffer_end;
 	uint8_t fifo_tag;
-	uint8_t tot_accel_fifo_words = 0;
+	uint16_t tot_accel_fifo_words = 0;
 	uint8_t tot_ts_fifo_words = 0;
 
 	buffer += sizeof(struct lis2dux12_fifo_data);
